+++ test/util/deleteAll.test.ts
@@ -1,7 +1,7 @@
 import { runDbMigrations } from '../../src/migrate';
 import Eval from '../../src/models/eval';
 import { deleteAllEvals } from '../../src/util';
-import EvalFactory from './factories/evalFactory';
+import EvalFactory from '../factories/evalFactory';
 
 describe('delete all evals', () => {
   beforeAll(async () => {
